requirements.txt
fastapi==0.73.0
uvicorn[standard]==0.17.4
tweepy==4.6.0
python-dotenv==0.19.2